Added drug update and drug sales

+++ src/main/java/com/bexos/PharmacyManagementSystem/dtos/DrugDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 public record DrugDto(
         String name,
         String description,
-        int price,
+        double price,
         int quantity,
         String supplier
 ) {
